Some refactoring (additional folders and files location changed) + dialogs service added

+++ src/lib/viewmodels/achievements/achievement_details_viewmodel.dart
@@ -1,16 +1,18 @@
 import 'dart:async';
 
 import 'package:event_bus/event_bus.dart';
-import 'package:kudosapp/models/achievement_holder.dart';
+import 'package:flutter/material.dart';
+import 'package:kudosapp/dto/achievement_holder.dart';
+import 'package:kudosapp/dto/user.dart';
+import 'package:kudosapp/models/achievement_model.dart';
 import 'package:kudosapp/models/achievement_to_send.dart';
 import 'package:kudosapp/models/list_notifier.dart';
 import 'package:kudosapp/models/messages/achievement_updated_message.dart';
-import 'package:kudosapp/models/user.dart';
 import 'package:kudosapp/service_locator.dart';
-import 'package:kudosapp/services/achievements_service.dart';
 import 'package:kudosapp/services/base_auth_service.dart';
-import 'package:kudosapp/services/people_service.dart';
-import 'package:kudosapp/viewmodels/achievement_viewmodel.dart';
+import 'package:kudosapp/services/database/achievements_service.dart';
+import 'package:kudosapp/services/database/people_service.dart';
+import 'package:kudosapp/services/dialogs_service.dart';
 import 'package:kudosapp/viewmodels/base_viewmodel.dart';
 
 enum OwnerType { user, team }
@@ -20,10 +22,11 @@ class AchievementDetailsViewModel extends BaseViewModel {
   final _peopleService = locator<PeopleService>();
   final _eventBus = locator<EventBus>();
   final _authService = locator<BaseAuthService>();
+  final _dialogsService = locator<DialogsService>();
   final String _achievementId;
 
   StreamSubscription<AchievementUpdatedMessage> _subscription;
-  AchievementViewModel _achievementViewModel;
+  AchievementModel _achievementViewModel;
   double _statisticsValue = 0;
   String _ownerName = "";
   String _ownerId = "";
@@ -31,7 +34,7 @@ class AchievementDetailsViewModel extends BaseViewModel {
 
   final achievementHolders = new ListNotifier<AchievementHolder>();
 
-  AchievementViewModel get achievementViewModel => _achievementViewModel;
+  AchievementModel get achievementViewModel => _achievementViewModel;
 
   double get statisticsValue => _statisticsValue;
 
@@ -56,7 +59,7 @@ class AchievementDetailsViewModel extends BaseViewModel {
     final achievement =
         await _achievementsService.getAchievement(_achievementId);
 
-    _achievementViewModel = AchievementViewModel(achievement);
+    _achievementViewModel = AchievementModel(achievement);
     _subscription?.cancel();
     _subscription =
         _eventBus.on<AchievementUpdatedMessage>().listen(_onAchievementUpdated);
@@ -84,14 +87,20 @@ class AchievementDetailsViewModel extends BaseViewModel {
     isBusy = false;
   }
 
-  Future<void> delete() async {
-    isBusy = true;
+  Future<void> deleteAchievement(BuildContext context) async {
+    if (await _dialogsService.showDeleteCancelDialog(
+        context: context,
+        title: localizer().warning,
+        content: localizer().deleteAchievementWarning)) {
+      isBusy = true;
 
-    await _achievementsService.deleteAchievement(
-        achievementViewModel.achievement.id,
-        holdersCount: achievementHolders.items?.length ?? 0);
+      await _achievementsService.deleteAchievement(
+          achievementViewModel.achievement.id,
+          holdersCount: achievementHolders.items?.length ?? 0);
 
-    isBusy = false;
+      isBusy = false;
+      Navigator.popUntil(context, ModalRoute.withName('/'));
+    }
   }
 
   @override
